sg: include MSP toolchain in default builds
